comment edit

+++ src/store/usePost.ts
@@ -2,17 +2,23 @@ import React from "react";
 import { useParams } from "react-router-dom";
 import { firestore } from "../config/firebase";
 import { Post } from "../types";
-export default function usePost(listen: Boolean = false): [Post] {
+export default function usePost(
+  listen: Boolean = false
+): [Post | null | undefined] {
   const { post_id: id } = useParams<{ post_id: string }>();
-  const [item, setItem] = React.useState<Post>({} as Post);
+  const [item, setItem] = React.useState<Post | null | undefined>(undefined);
   React.useEffect(() => {
     if (listen) {
       return firestore
         .collection("posts")
         .doc(id)
         .onSnapshot((doc) => {
-          const item = { id: doc.id, ...doc.data() } as Post;
-          setItem(item);
+          if (doc.exists) {
+            const item = { id: doc.id, ...doc.data() } as Post;
+            setItem(item);
+          } else {
+            setItem(null);
+          }
         });
     } else {
       firestore
@@ -20,8 +26,12 @@ export default function usePost(listen: Boolean = false): [Post] {
         .doc(id)
         .get()
         .then((doc) => {
-          const item = { id: doc.id, ...doc.data() } as Post;
-          setItem(item);
+          if (doc.exists) {
+            const item = { id: doc.id, ...doc.data() } as Post;
+            setItem(item);
+          } else {
+            setItem(null);
+          }
         });
     }
   }, [id, listen]);
